♻️ refactor: refactor the aiChat store
