+++ src/features/ChatInput/useSend.ts
@@ -2,8 +2,8 @@ import { useCallback, useMemo } from 'react';
 
 import { useChatStore } from '@/store/chat';
 import { chatSelectors } from '@/store/chat/selectors';
-import { SendMessageParams } from '@/store/chat/slices/aiChat/action';
 import { fileChatSelectors, useFileStore } from '@/store/file';
+import { SendMessageParams } from '@/types/message';
 
 export type UseSendMessageParams = Pick<
   SendMessageParams,
